update preceptor-related function and change the function folder name from lib to functions
