+++ functions/function1/.prettierrc.json
