+++ src/components/ui/checkbox.jsx
@@ -9,7 +9,7 @@ import * as React from "react"
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox"
 import { CheckIcon } from "lucide-react"
 
-import { cn } from "../../../lib/utils"
+import { cn } from "../../../functions/utils"
 
 /**
  * @function Checkbox
